ci: disable code signing for third-party pull requests
Third-party pull requests won't have access to the repository secrets
needed to perform code signing, so we'll just perform an unsigned build
if those secrets aren't available.

+++ scripts/ci/build.sh
@@ -8,31 +8,43 @@ MUTAGEN_OS_NAME="$(go env GOOS)"
 
 # Perform a build that's appropriate for the platform.
 if [[ "${MUTAGEN_OS_NAME}" == "darwin" ]]; then
-    # Compute the path and password for a temporary keychain where we'll import
-    # the macOS code signing certificate and private key.
-    MUTAGEN_KEYCHAIN_PATH="${RUNNER_TEMP}/mutagen.keychain-db"
-    MUTAGEN_KEYCHAIN_PASSWORD="$(dd if=/dev/random bs=1024 count=1 2>/dev/null | openssl dgst -sha256)"
+    # Check if code signing is possible. If so, then set up the keychain and
+    # certificates that we'll need and perform a signed build. If not, then just
+    # perform a normal build.
+    if [[ ! -z "${MACOS_CODESIGN_IDENTITY}" ]]; then
+        # Compute the path and password for a temporary keychain where we'll import
+        # the macOS code signing certificate and private key.
+        MUTAGEN_KEYCHAIN_PATH="${RUNNER_TEMP}/mutagen.keychain-db"
+        MUTAGEN_KEYCHAIN_PASSWORD="$(dd if=/dev/random bs=1024 count=1 2>/dev/null | openssl dgst -sha256)"
 
-    # Store the previous default keychain.
-    PREVIOUS_DEFAULT_KEYCHAIN="$(security default-keychain | xargs)"
+        # Store the previous default keychain.
+        PREVIOUS_DEFAULT_KEYCHAIN="$(security default-keychain | xargs)"
 
-    # Create the temporary keychain, set it to be the default keychain, set it
-    # to automatically re-lock (just in case removal fails), and unlock it.
-    security create-keychain -p "${MUTAGEN_KEYCHAIN_PASSWORD}" "${MUTAGEN_KEYCHAIN_PATH}"
-    security default-keychain -s "${MUTAGEN_KEYCHAIN_PATH}"
-    security set-keychain-settings -lut 3600 "${MUTAGEN_KEYCHAIN_PATH}"
-    security unlock-keychain -p "${MUTAGEN_KEYCHAIN_PASSWORD}" "${MUTAGEN_KEYCHAIN_PATH}"
+        # Create the temporary keychain, set it to be the default keychain, set it
+        # to automatically re-lock (just in case removal fails), and unlock it.
+        security create-keychain -p "${MUTAGEN_KEYCHAIN_PASSWORD}" "${MUTAGEN_KEYCHAIN_PATH}"
+        security default-keychain -s "${MUTAGEN_KEYCHAIN_PATH}"
+        security set-keychain-settings -lut 3600 "${MUTAGEN_KEYCHAIN_PATH}"
+        security unlock-keychain -p "${MUTAGEN_KEYCHAIN_PASSWORD}" "${MUTAGEN_KEYCHAIN_PATH}"
 
-    # Import the macOS code signing certificate and private key and allow access
-    # from the codesign utility.
-    MUTAGEN_CERTIFICATE_AND_KEY_PATH="${RUNNER_TEMP}/certificate_and_key.p12"
-    echo -n "${MACOS_CODESIGN_CERTIFICATE_AND_KEY}" | base64 --decode --output "${MUTAGEN_CERTIFICATE_AND_KEY_PATH}"
-    security import "${MUTAGEN_CERTIFICATE_AND_KEY_PATH}" -k "${MUTAGEN_KEYCHAIN_PATH}" -P "${MACOS_CODESIGN_CERTIFICATE_AND_KEY_PASSWORD}" -T "/usr/bin/codesign"
-    rm "${MUTAGEN_CERTIFICATE_AND_KEY_PATH}"
-    security set-key-partition-list -S apple-tool:,apple: -s -k "${MUTAGEN_KEYCHAIN_PASSWORD}" "${MUTAGEN_KEYCHAIN_PATH}" > /dev/null
+        # Import the macOS code signing certificate and private key and allow access
+        # from the codesign utility.
+        MUTAGEN_CERTIFICATE_AND_KEY_PATH="${RUNNER_TEMP}/certificate_and_key.p12"
+        echo -n "${MACOS_CODESIGN_CERTIFICATE_AND_KEY}" | base64 --decode --output "${MUTAGEN_CERTIFICATE_AND_KEY_PATH}"
+        security import "${MUTAGEN_CERTIFICATE_AND_KEY_PATH}" -k "${MUTAGEN_KEYCHAIN_PATH}" -P "${MACOS_CODESIGN_CERTIFICATE_AND_KEY_PASSWORD}" -T "/usr/bin/codesign"
+        rm "${MUTAGEN_CERTIFICATE_AND_KEY_PATH}"
+        security set-key-partition-list -S apple-tool:,apple: -s -k "${MUTAGEN_KEYCHAIN_PASSWORD}" "${MUTAGEN_KEYCHAIN_PATH}" > /dev/null
 
-    # Perform a full release build.
-    go run scripts/build.go --mode=release --macos-codesign-identity="${MACOS_CODESIGN_IDENTITY}"
+        # Perform a full release build with code signing.
+        go run scripts/build.go --mode=release --macos-codesign-identity="${MACOS_CODESIGN_IDENTITY}"
+
+        # Reset the default keychain and remove the temporary keychain.
+        security default-keychain -s "${PREVIOUS_DEFAULT_KEYCHAIN}"
+        security delete-keychain "${MUTAGEN_KEYCHAIN_PATH}"
+    else
+        # Perform a full release build without code signing.
+        go run scripts/build.go --mode=release
+    fi
 
     # Determine the Mutagen version.
     MUTAGEN_VERSION="$(build/mutagen version)"
@@ -56,11 +68,8 @@ if [[ "${MUTAGEN_OS_NAME}" == "darwin" ]]; then
     tar xzf "build/release/mutagen_windows_arm64_v${MUTAGEN_VERSION}.tar.gz"
     zip "build/release/mutagen_windows_arm64_v${MUTAGEN_VERSION}.zip" mutagen.exe mutagen-agents.tar.gz
     rm mutagen.exe mutagen-agents.tar.gz
-
-    # Reset the default keychain and remove the temporary keychain.
-    security default-keychain -s "${PREVIOUS_DEFAULT_KEYCHAIN}"
-    security delete-keychain "${MUTAGEN_KEYCHAIN_PATH}"
 else
+    # Perform a slim build.
     go run scripts/build.go --mode=slim
 fi
